+++ src/components/ContextMenu/ContextMenu.tsx
@@ -1,0 +1,49 @@
+"use client";
+
+import React from "react";
+import DropdownMenu from "components/DropdownMenu";
+import useScrollLock from "hooks/useScrollLock";
+import type * as G from "types/global";
+import type * as T from "./ContextMenu.types";
+import s from "./ContextMenu.module.css";
+
+const ContextMenu = (props: T.Props) => {
+	const { position = "end-top", ...dropdownMenuProps } = props;
+	const [coordinates, setCoordinates] = React.useState<G.Coordinates>();
+	const containerRef = React.useRef<HTMLDivElement>(null);
+	const { lockScroll, unlockScroll } = useScrollLock({ containerRef });
+
+	React.useEffect(() => {
+		const handleContextMenu = (e: MouseEvent) => {
+			e.preventDefault();
+			setCoordinates({ x: e.clientX, y: e.clientY });
+			lockScroll();
+		};
+
+		window.addEventListener("contextmenu", handleContextMenu);
+		return () => window.removeEventListener("contextmenu", handleContextMenu);
+	}, [lockScroll]);
+
+	return (
+		<div className={s.root} ref={containerRef}>
+			<DropdownMenu
+				{...dropdownMenuProps}
+				position={position}
+				originCoordinates={coordinates}
+				active={!!coordinates}
+				onClose={() => {
+					setCoordinates(undefined);
+					unlockScroll();
+				}}
+			/>
+		</div>
+	);
+};
+
+ContextMenu.Content = DropdownMenu.Content;
+ContextMenu.Item = DropdownMenu.Item;
+ContextMenu.Section = DropdownMenu.Section;
+ContextMenu.SubMenu = DropdownMenu.SubMenu;
+ContextMenu.SubTrigger = DropdownMenu.SubTrigger;
+
+export default ContextMenu;