+++ src/components/ContextMenu/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./ContextMenu";
+export type { Props as ContextMenuProps } from "./ContextMenu.types";