+++ src/components/Tooltip/Tooltip.tsx
@@ -7,38 +7,12 @@ import type * as T from "./Tooltip.types";
 import s from "./Tooltip.module.css";
 
 const Tooltip = (props: T.Props) => {
-	const {
-		id,
-		text,
-		children,
-		onOpen,
-		onClose,
-		position = "bottom",
-		containerRef,
-		contentGap,
-		contentShift,
-		active,
-		disabled,
-		disableContentHover,
-	} = props;
+	const { text, children, position = "bottom", ...flyoutProps } = props;
 
 	if (!text) return children({});
 
 	return (
-		<Flyout
-			id={id}
-			active={active}
-			position={position}
-			disabled={disabled}
-			contentGap={contentGap}
-			contentShift={contentShift}
-			onOpen={onOpen}
-			onClose={onClose}
-			disableContentHover={disableContentHover}
-			containerRef={containerRef}
-			triggerType="hover"
-			groupTimeouts
-		>
+		<Flyout {...flyoutProps} position={position} triggerType="hover" groupTimeouts>
 			<Flyout.Trigger>{children}</Flyout.Trigger>
 			<Flyout.Content>
 				<Theme colorMode="inverted">