+++ src/utilities/dom/flyout.ts
@@ -1,22 +1,16 @@
-import { getShadowRoot } from "./shadowDom";
+import type * as G from "types/global";
 
-export const getClosestFlyoutTarget = (
-	el: HTMLElement | null,
-	iteration: number = 0
-): HTMLElement => {
-	const style = el && window.getComputedStyle(el);
-	const overflowY = style?.overflowY;
-	const position = style?.position;
-	const isScrollable = overflowY?.includes("scroll");
-	const isFixed = position === "fixed" || position === "sticky";
+export const getRectFromCoordinates = (coordinates: DOMRect | G.Coordinates): DOMRect => {
+	if ("width" in coordinates && coordinates.width !== undefined) return coordinates;
 
-	// Only check shadow root on the first run
-	if (iteration === 0) {
-		const shadowRoot = getShadowRoot(el);
-		if (shadowRoot?.firstElementChild) return shadowRoot.firstElementChild as HTMLElement;
-	}
-
-	if (el === document.body || !el) return document.body;
-	if ((isScrollable && el.scrollHeight > el.clientHeight) || isFixed) return el;
-	return getClosestFlyoutTarget(el.parentElement, iteration + 1);
+	return {
+		...coordinates,
+		width: 0,
+		height: 0,
+		left: coordinates.x,
+		right: coordinates.x,
+		top: coordinates.y,
+		bottom: coordinates.y,
+		toJSON: () => {},
+	};
 };