+++ src/utilities/dom/index.ts
@@ -1,4 +1,4 @@
-export { getClosestFlyoutTarget } from "./flyout";
+export { getRectFromCoordinates } from "./flyout";
 export { getShadowRoot } from "./shadowDom";
 export { enableUserSelect, disableUserSelect } from "./userSelect";
-export { findParent } from "./find";
+export { findParent, findClosestRenderContainer } from "./find";