+++ src/utilities/scroll/lock.ts
@@ -1,6 +1,7 @@
 import { isIOS } from "utilities/platform";
 import lockSafariScroll from "./lockSafari";
 import lockStandardScroll from "./lockStandard";
+import { findClosestRenderContainer } from "utilities/dom";
 
 let lockedCount = 0;
 let reset = () => {};
@@ -10,10 +11,14 @@ export const lockScroll = (args: { containerEl?: HTMLElement | null; cb?: () => 
 
 	if (lockedCount > 1) return;
 
+	const customLockTargetEl = args.containerEl
+		? findClosestRenderContainer({ el: args.containerEl, overflowOnly: true })
+		: undefined;
+
 	if (isIOS()) {
-		reset = lockSafariScroll();
+		reset = lockSafariScroll(customLockTargetEl);
 	} else {
-		reset = lockStandardScroll(args.containerEl);
+		reset = lockStandardScroll(customLockTargetEl);
 	}
 
 	args.cb?.();